+++ changelog_entry.yaml
@@ -1,4 +1,4 @@
 - bump: minor
   changes:
     added: 
-    - Alberta marital status credit.
+    - Alberta spouse and common-law partner amount.
